Split PDG powered descent solver into partial class
